requirements.txt
sentence-transformers
pinecone
llama-cpp-python
python-dotenv
